refactor: roomForm component
